+++ Application/DTOs/SortFilterDTOs/CompanySortFilterDto.cs
@@ -4,6 +4,6 @@ namespace Application.DTOs.SortFilterDTOs;
 
 public class CompanySortFilterDto : BaseSortFilterDto
 {
-    public IList<long> IdList { get; set; } = [];
-    public IList<long> CountryIdList { get; set; } = [];
+    public IList<int> IdList { get; set; } = [];
+    public IList<int> CountryIdList { get; set; } = [];
 }
