+++ Application/DTOs/SortFilterDTOs/LocationSortFilterDto.cs
@@ -4,6 +4,6 @@ namespace Application.DTOs.SortFilterDTOs;
 
 public class LocationSortFilterDto : BaseSortFilterDto
 {
-    public IList<long> IdList { get; set; } = [];
-    public long CountryId { get; set; }
+    public IList<int> IdList { get; set; } = [];
+    public int CountryId { get; set; }
 }
