+++ Application/DTOs/SortFilterDTOs/UserSortFilterDto.cs
@@ -4,7 +4,7 @@ namespace Application.DTOs.SortFilterDTOs;
 
 public class UserSortFilterDto : BaseSortFilterDto
 {
-    public IList<long> IdList { get; set; } = [];
-    public IList<long> LocationIdList { get; set; } = [];
-    public IList<long> CategoryIdList { get; set; } = [];
+    public IList<int> IdList { get; set; } = [];
+    public IList<int> LocationIdList { get; set; } = [];
+    public IList<int> CategoryIdList { get; set; } = [];
 }
