+++ Application/Services/Search/Application/ApplicationSearchModel.cs
@@ -5,7 +5,7 @@ namespace Application.Services.Search.Application;
 
 public class ApplicationSearchModel : BaseSearchModel
 {
-    public long JobId { get; set; }
+    public int JobId { get; set; }
     public IList<EducationRecord> EducationList { get; set; } = [];
     public IList<WorkRecord> WorkRecordList { get; set; } = [];
     public IList<string> SkillList { get; set; } = [];
