+++ Application/Services/Search/Location/LocationSearchModel.cs
@@ -4,7 +4,7 @@ namespace Application.Services.Search.Location;
 
 public class LocationSearchModel : BaseSearchModel
 {
-    public long CountryId { get; set; }
+    public int CountryId { get; set; }
     public string Name { get; set; } = "";
     public string? Description { get; set; }
     public string? Code { get; set; }
