+++ Application/Services/Search/User/IUserSearchService.cs
@@ -4,5 +4,5 @@ namespace Application.Services.Search.User;
 
 public interface IUserSearchService : IBaseSearchService<UserSearchModel>
 {
-    Task<IList<long>> SearchForCountryId(string query, long countryId);
+    Task<IList<int>> SearchForCountryId(string query, int countryId);
 }
