+++ Domain/Common/ITreeEntity.cs
@@ -2,6 +2,6 @@
 
 public interface ITreeEntity
 {
-    public long? ParentId { get; set; }
+    public int? ParentId { get; set; }
     public int Level { get; set; }
 }
