+++ Domain/Entities/UserCompanyBookmark.cs
@@ -5,9 +5,8 @@ namespace Domain.Entities;
 public class UserCompanyBookmark : BaseEntity
 {
     public virtual User? User { get; set; }
-    public long UserId { get; set; }
+    public int UserId { get; set; }
     
     public virtual Company? Company { get; set; }
-    public long CompanyId { get; set; }
-
+    public int CompanyId { get; set; }
 }
