+++ Domain/Entities/UserJobBookmark.cs
@@ -5,8 +5,8 @@ namespace Domain.Entities;
 public class UserJobBookmark : BaseEntity
 {
     public virtual User? User { get; set; }
-    public long UserId { get; set; }
+    public int UserId { get; set; }
     
     public virtual Job? Job { get; set; }
-    public long JobId { get; set; }
+    public int JobId { get; set; }
 }
